requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
SQLAlchemy==1.3.10
mysqlclient==1.3.14
pyodbc==4.0.27
requests-oauthlib==1.3.0
gunicorn